Refactoring to decuple the vectorization process
