+++ app/Services/DocumentManager.php
@@ -2,21 +2,24 @@
 
 namespace App\Services;
 
+use App\Models\Document;
+use PhpParser\Comment\Doc;
 use App\Enums\DocumentType;
 use App\Jobs\ProcessPDFDocument;
-use App\Jobs\ProcessWebPageDocument;
-use App\Models\Document;
+use App\Jobs\ProcessTextDocument;
 use Illuminate\Support\Collection;
+use App\Jobs\ProcessWebPageDocument;
 
 class DocumentManager
 {
     public function process(Document $document): void
     {
-        if ($document->type === DocumentType::PDF) {
-            ProcessPDFDocument::dispatch($document);
-        } elseif ($document->type === DocumentType::WEB_PAGE) {
-            ProcessWebPageDocument::dispatch($document);
-        }
+        match ($document->type) {
+            DocumentType::PDF => ProcessPDFDocument::dispatch($document),
+            DocumentType::WEB_PAGE => ProcessWebPageDocument::dispatch($document),
+            DocumentType::RAW_TEXT =>  ProcessTextDocument::dispatch($document),
+            default => throw new \Exception('Unsupported document type'),
+        };
     }
 
     public function chunk(Document $document): Collection
